move argument parsing into own file and make injectScrips.js executable as a script for single file
